requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.23
asyncpg==0.29.0        # Для PostgreSQL (на будущее)
aiosqlite==0.20.0      # Для SQLite
redis==5.0.1
python-dotenv==1.0.0
apscheduler==3.10.4